requirements.txt
Flask
Flask-Bcrypt
Flask-DebugToolbar
Flask-SQLAlchemy
Flask-WTF
ipython
psycopg2-binary
email_validatorappdirs==1.4.4
backcall==0.2.0
bcrypt==3.2.0
black==21.7b0
blinker==1.4
cffi==1.14.6
click==8.0.1
decorator==5.0.9
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-DebugToolbar==0.11.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.0
gunicorn==20.1.0
idna==3.2
ipython==7.25.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
ipython-genutils==0.2.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mypy-extensions==0.4.3
parso==0.8.2
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.19
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.9.0
regex==2021.7.6
six==1.16.0
SQLAlchemy==1.4.21
tomli==1.0.4
traitlets==5.0.5
wcwidth==0.2.5
Werkzeug==2.0.1
WTForms==2.3.3
